requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
sqlalchemy==2.0.36
apscheduler==3.10.4
twilio==9.3.7
python-telegram-bot==21.7
pydantic==2.10.3
python-dateutil==2.9.0.post0
